fix(executor): use u32 for ColumnSource::table_oid to match catalog types
TableInfo::table_id and Plan::SeqScan::table_id are both u32, but
ColumnSource::table_oid was i32, requiring a lossy `as i32` cast in
the planner. Move the i32 conversion to the wire protocol boundary
(connection.rs) where PostgreSQL's signed OID format is needed.

+++ src/executor/planner.rs
@@ -159,7 +159,7 @@ fn build_column_descs(
             name: col.column_name.clone(),
             source: Some(ColumnSource {
                 table_name: table_name.to_string(),
-                table_oid: table_id as i32,
+                table_oid: table_id,
                 column_id: (i + 1) as i16,
             }),
             data_type: col.data_type,
